generic: change signature of print_hex to use const uint8_t
The argument to print_hex should not be modified, therefore it should be
const so that other const strings need not discard their const via cast.

While doing so, change unsigned char to uint8_t, which is functionally
the same, but shorter and more accurate to what is being printed.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>