guest/read: return early if next esd cannot be parsed
Matching the rest of the function, don't continue iterating through
ESDs if one fails to parse.

Also adjusts the error message on signature type to state "unknown"
signature type instead of invalid, since it may not be supported yet.
This is the one "error" case that continues without exiting early.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>