guest/generate: reword error message for disallowing PK appends
There can only be one PK, therefore the PK cannot be appended to.
However, the current error message is not clear, and references the
older method of setting the flag. Furthermore, this error is handled in
the same way as an invalid argument parse which leads to the whole usage
information being printed, which makes finding the error message
difficult.

This commit rewords the error message to be a little more clear, and
moves the special case check for PK appends outside of the rest of the
argument validation logic, thus avoiding the extra usage output.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>