host/edk2: use struct efi_time[] as the last timestamp type instead of char*
Many of the timestamp-related functions in the host edk2 backend take
in both a single timestamp and a reference to the contents of the TS
variable, which is really an array of timestamps.

This is likely why the the type of that array was left as char*, so it
wouldn't be ambiguous with the single `struct efi_time*` input
parameters. This commit changes the char* usage to be
`struct efi_time[]`, which conveys the type information while still
being distinct from the single `struct efi_time*` inputs.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>