guest: add timestamp print output to read/verify commands
Move the existing print_timestamp function to common/util, and add a new
helper to handle reading and printing a timestamp from a variable buffer.

Prints the timestamp when using the read command, and when using verbose
levels in the verify command.

This commit is part of a larger work in progress on factoring out
loading data from a variable vs a file, see more information in #76.

Co-developed-by: Eric Richter <erichte@linux.ibm.com>
Signed-off-by: Sudhakar Kuppusamy <sudhakar@linux.ibm.com>