external/skiboot: commented out global extern definitions in secvar.h
The globals used in skiboot proper are not used in secvarctl, thus do
not need to be included in this header. These globals are then shadowed
in the host backend when simulating updates, causing warnings from
cppcheck.

Rather than suppress the warnings, comment out the offending extern
declarations, which should probably not be there in the first place.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>