guest: replace TIMESTAMP_LEN with actual sized structs
TIMESTAMP_LEN is incredibly misleading. In its current form, it refers
to an 8-byte header consisting of a 1-byte version number, and 7 bytes of
a timestamp (truncated at pad1). This is obviously confusing as
sizeof(timestamp_t) == 16, and this also includes a version byte for some
reason.

This commit attempts to clean this up a bit, by using the struct defined
in libstb-secvar that defines the layout of this header, so that types
and sizeof() can be used to get these offset/size values, instead of
undocumented mislabled constants.

This includes a version bump of libstb-secvar, which includes a commit
exposing the variable header struct in pseries.h.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>