generic: fix static builds by removing unnecessary header include
Introduced in 74c5b430 .

While reworking buffer types, a header include got added to the top of
generic.c (likely automatically placed by an editor). This header does
not seem to be present in the alpine build environment, thus causing the
static builds to fail.

Since this header is not even needed, it can be removed.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>