guest/generate: fix PK append flag error being printed multiple times
Argp calls the `parse_options()` function once per argument, therefore it
is possible that the condition for printing the error may be reached if
there are more arguments to parse.

Move the check to after arguments have been parsed, so that the check is
only made after the complete argument stream has been parsed in entirety.

Fixes: #80

Reported-by: R Nageswara Sastry <rnsastry@linux.ibm.com>
Signed-off-by: Eric Richter <erichte@linux.ibm.com>