guest/read: don't attempt to print an ESL from a zero-sized buffer
Currently, print_esl_buffer() passes along the input buffer and size to
next_esl_from_buffer(), which will immediately error since a buffer size
of zero obviously cannot contain an ESL.

This can sometimes happen when printing a variable that contains a
timestamp but no data -- the most obvious example being a variable that
was cleared.

Therefore, a message should be printed to indicate there is no data,
rather than an error message that should be reserved for when the data
is actually malformed.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>