test/guest: use named variables instead of array of directories, use os.path.join
An array of paths is _very_ unclear when reading a section of code
without context. Use named globals instead so that the reader does not
need to move around the file to translate.

Also use os.path.join() everywhere instead of manually joining paths with
an add operation. This ensures path separators are included where needed,
and also helps with portability.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>