test/guest: merge variable generation lists to remove duplicate code
There is no need to separate the list of variables signed by PK or by KEK,
the third element of that list does that for us. Merging them, and using
list decompositon, we can simplify the generation functions to use one
loop and avoid list access by semi-arbitrary index numbers.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>